+++ apps/commerce-api/src/main/java/com/loopers/domain/like/brand/BrandLikeRepository.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.like;
+package com.loopers.domain.like.brand;
 
 import java.util.Optional;
