+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/like/BrandLikeJpaRepository.java
@@ -1,6 +1,6 @@
 package com.loopers.infrastructure.like;
 
-import com.loopers.domain.like.BrandLikeModel;
+import com.loopers.domain.like.brand.BrandLikeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
