+++ apps/commerce-api/src/test/java/com/loopers/application/like/BrandLikeFacadeIntegrationTest.java
@@ -3,8 +3,8 @@ package com.loopers.application.like;
 import com.loopers.domain.brand.BrandFixture;
 import com.loopers.domain.brand.BrandModel;
 import com.loopers.domain.brand.BrandRepository;
-import com.loopers.domain.like.BrandLikeModel;
-import com.loopers.domain.like.BrandLikeRepository;
+import com.loopers.domain.like.brand.BrandLikeModel;
+import com.loopers.domain.like.brand.BrandLikeRepository;
 import com.loopers.domain.like.fixture.BrandLikeFixture;
 import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
