+++ apps/commerce-api/src/test/java/com/loopers/application/like/BrandLikeFacadeTest.java
@@ -4,9 +4,9 @@ import com.loopers.application.brand.BrandFacade;
 import com.loopers.domain.brand.BrandModel;
 import com.loopers.domain.brand.BrandRepository;
 import com.loopers.domain.brand.BrandFixture;
-import com.loopers.domain.like.BrandLikeModel;
-import com.loopers.domain.like.BrandLikeRepository;
-import com.loopers.domain.like.BrandLikeService;
+import com.loopers.domain.like.brand.BrandLikeModel;
+import com.loopers.domain.like.brand.BrandLikeRepository;
+import com.loopers.domain.like.brand.BrandLikeService;
 import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
 import org.junit.jupiter.api.DisplayName;
